agent-server/pyproject.toml
[project]
name = "msme-agent-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "certifi>=2026.1.4",
    "fastapi>=0.128.0",
    "langchain>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.7",
    "motor>=3.7.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
    "websockets>=15.0.1",
]
